LayoutTests/platform/ios-simulator/svg/css/parse-height-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="auto"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="  100"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="100   "
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="100   px"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="100px;"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="100px !important"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="{ 100px }"
CONSOLE MESSAGE: Error: Invalid negative value for <rect> attribute height="-100px"
Test that 'with' presentation attribute is parsed with CSS presentation rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle("height", "auto") is "auto"
PASS computedStyle("height", "  100") is "100px"
PASS computedStyle("height", "100   ") is "100px"
PASS computedStyle("height", "100px") is "100px"
PASS computedStyle("height", "1em") is "16px"
FAIL computedStyle("height", "1ex") should be 12.800000190734863px. Was 16px.
PASS computedStyle("height", "20%") is "20%"
PASS computedStyle("height", "100   px") is "auto"
PASS computedStyle("height", "100px;") is "auto"
PASS computedStyle("height", "100px !important") is "auto"
PASS computedStyle("height", "{ 100px }") is "auto"
PASS computedStyle("height", "-100px") is "auto"
PASS successfullyParsed is true

TEST COMPLETE
